refactor: add Redux to handle state changes of the app
Until now we used stateful React components (eg. the Editor organism) to handle state changes
(through `setState()`). However, this method is not scalable. In the upcoming commits we'll have to
add more states to the app, such as async code validation. For this reason we add Redux state
management to this project.